requirements.txt
aiobotocore==2.24.2
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.17.0
amazon-transcribe==0.6.4
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
attrs==25.3.0
awscrt==0.26.1
boto3==1.40.18
botocore==1.40.18
caio==0.9.24
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.2
ecdsa==0.19.1
fastapi==0.104.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.6.4
propcache==0.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
